Add ability to specify target for Microkit tool
Add the `--tool-target-triple` option to specify which target
to build the tool for. Unfortunately Rust does not allow you to
cross-compile from Linux to macOS, however this will allow x64 macOS
hosts to cross-compile to ARM64 macOS. In theory the reverse would
work as well but I have not tried it.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>