loader: only use general registers
There is a naive memcpy() implementation in the loader
for copying data from the loader image to the right
region in physical memory. The compiler, understandibly, decides
to optimise this (given we are passing -O3). However, the
compiler optimizes this loop into vector/FPU instructions
which I believe fails on platforms that do not already have
their FPU initialised, such as QEMU ARM virt.

Most hardware platforms have their FPU initialised by U-Boot
which runs before the Microkit loader which is why it has
not been a problem until now.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>