Increase loader link address for MaaXBoard
It was seen on larger Microkit systems that the loader data was
overlapping with the code for the loader itself when the loader
attempted to copy the data to the appropriate location before
jumping to seL4 and starting the system.

I believe the root cause of this problem is not gone and will involve
changing the tool to be smarter with how it allocates physical memory,
my current attempts to do so lead to other issues. I'll have to make a
follow up fix to prevent this from happening again, but for now this
will allow users of the MaaXBoard to run larger Microkit systems.

Signed-off-by: Ivan Velickovic <i.velickovic@unsw.edu.au>