tool: specify ppcalls on a per-channel basis
Previously, this was done on a per-protection domain basis, which
meant that protection domains were given more rights than was strictly
necessary. This also allows for authors to remove rights for sending
notifications on either end of the channel, and so allows for having
unidirectional notification channels.

This removes the previous `pp="true"` attribute on PDs (the one you
would PPC *into*), replacing it with `pp="true/false" and
`notify="true/false"` on each channel end.

So the following:
```
<protection_domain name="sender" [...] />
<protection_domain name="receiver" pp="true" [...] />
<channel>
  <end id="0" pd="sender"   />
  <end id="0" pd="receiver" />
</channel>
```

Becomes instead:
```
<protection_domain name="sender" [...] />
<protection_domain name="receiver" [...] />
<channel>
  <end id="0" pd="sender" pp="true"  />
  <end id="0" pd="receiver"          />
</channel>
```

Signed-off-by: julia <git.ts@trainwit.ch>